sending opponent's name to view
