+++ Client/src/main/resources/templates/result.html
@@ -19,7 +19,7 @@
 </script>
 
 <script type="text/javascript">
-    setInterval(refreshName, 500);
+    setInterval(refreshName, 5000);
 </script>
 
 <body>
